requirements.txt
Babel==2.4.0
Django==1.11.4
Django-Select2==5.11.1
Faker==0.7.18
Pillow==4.2.1
Pygments==2.2.0
Unidecode==0.04.21
amqp==2.2.1
appdirs==1.4.3
awesome-slugify==1.6.5
backports-abc==0.5
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
decorator==4.1.2
django-appconf==1.0.2
git+https://github.com/kostko/django-autoslug.git@fix/empty-slug
django-bootstrap3==9.0.0
django-braces==1.11.0
django-celery-email==2.0.0
django-ckeditor==5.3.0
django-compressor==2.1.1
django-countries==4.6.1
django-crispy-forms==1.6.1
django-debug-panel==0.8.3
django-debug-toolbar==1.8
django-extensions==1.8.1
django-grappelli==2.10.1
django-htmlmin==0.10.0
django-imagekit==4.0.1
django-js-asset==0.1.1
django-model-utils==3.0.0
django-recaptcha==1.3.1
django-redis==4.8.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
flower==0.9.2
gunicorn==19.7.1
html5lib==0.999999999
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
kombu==4.1.0
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
rcssmin==1.0.6
redis==2.10.5
regex==2017.07.28
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
tornado==4.5.1
traitlets==4.3.2
typing==3.6.1
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
bs4==0.0.1
mutagen==1.38
requests==2.18.4
django-cleanup==1.0.1
textile==2.3.16
social-auth-app-django==1.2.0
social-auth-core==1.4.0